# docs/swagger.yaml
basePath: /
definitions:
  entity.APIToken:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      name:
        type: string
      token:
        type: string
      type:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
  entity.Attendance:
    properties:
      at:
        description: Time when attendance was taken
        type: string
      created_at:
        type: string
      head:
        $ref: '#/definitions/entity.User'
      head_id:
        description: The head who recorded this attendance
        type: integer
      id:
        type: integer
      session:
        $ref: '#/definitions/entity.Session'
      session_id:
        type: integer
      status:
        description: e.g., 0=absent, 1=present, 2=excused
        type: integer
      type:
        description: Type of attendance record
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
  entity.Comment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      post:
        $ref: '#/definitions/entity.Post'
      post_id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      replies:
        description: Relations
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      reply:
        $ref: '#/definitions/entity.Comment'
      reply_id:
        type: integer
      submission:
        $ref: '#/definitions/entity.Submission'
      submission_id:
        type: integer
      text:
        type: string
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
...
  entity.Country:
    properties:
      created_at:
        type: string
      groups:
        items:
          $ref: '#/definitions/entity.Group'
        type: array
      id:
        type: integer
      name:
        type: string
      short_code:
        type: string
      updated_at:
        type: string
      users:
        description: Relations
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
  entity.DailyProblem:
    properties:
      created_at:
        type: string
      for_date:
        type: string
      id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      super_group:
        $ref: '#/definitions/entity.SuperGroup'
      super_group_id:
        type: integer
      updated_at:
        type: string
    type: object
  entity.Exercise:
    properties:
      created_at:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      updated_at:
        type: string
    type: object
  entity.Fund:
    properties:
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      sessions:
        description: Relations
        items:
          $ref: '#/definitions/entity.Session'
        type: array
      stipends:
        items:
          $ref: '#/definitions/entity.Stipend'
        type: array
      updated_at:
        type: string
    type: object
  entity.GoogleOAuth:
    properties:
      calendar_id:
        type: string
      created_at:
        type: string
      encrypted_token_string:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
  entity.Group:
    properties:
      country:
        $ref: '#/definitions/entity.Country'
      country_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      exercises:
        items:
          $ref: '#/definitions/entity.Exercise'
        type: array
      google_oauths:
        items:
          $ref: '#/definitions/entity.GoogleOAuth'
        type: array
      group_sessions:
        items:
          $ref: '#/definitions/entity.GroupSession'
        type: array
      hoa_id:
        description: Head of Academy ID
        type: integer
      hoas:
        items:
          $ref: '#/definitions/entity.HOA'
        type: array
      id:
        type: integer
      invites:
        items:
          $ref: '#/definitions/entity.Invite'
        type: array
      name:
        type: string
      short_name:
        type: string
      super_to_groups:
        items:
          $ref: '#/definitions/entity.SuperToGroup'
        type: array
      updated_at:
        type: string
      users:
        description: Relations
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
  entity.GroupSession:
    properties:
      created_at:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      session:
        $ref: '#/definitions/entity.Session'
      session_id:
        type: integer
      updated_at:
        type: string
    type: object
  entity.HOA:
    properties:
      created_at:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
  entity.Invite:
    properties:
      created_at:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      key:
        description: Unique invite key
        type: string
      role:
        $ref: '#/definitions/entity.Role'
      role_id:
        type: integer
      updated_at:
        type: string
      used:
        description: Whether the invite has been used
        type: boolean
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        description: User who created the invite
        type: integer
    type: object
  entity.Post:
    properties:
      body:
        type: string
      comments:
        description: Relations
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      post_to_tags:
        items:
          $ref: '#/definitions/entity.PostToTag'
        type: array
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
...
  entity.PostToTag:
    properties:
      created_at:
        type: string
      id:
        type: integer
      post:
        $ref: '#/definitions/entity.Post'
      post_id:
        type: integer
      post_tag:
        $ref: '#/definitions/entity.PostTag'
      post_tag_id:
        type: integer
      updated_at:
        type: string
    type: object
  entity.Problem:
    properties:
      comments:
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      contest:
        $ref: '#/definitions/entity.Contest'
      contest_id:
        type: integer
      created_at:
        type: string
      daily_problems:
        items:
          $ref: '#/definitions/entity.DailyProblem'
        type: array
      difficulty:
        type: string
      exercises:
        items:
          $ref: '#/definitions/entity.Exercise'
        type: array
      id:
        type: integer
      link:
        type: string
      name:
        type: string
      platform:
        type: string
      problem_tracks:
        items:
          $ref: '#/definitions/entity.ProblemTrack'
        type: array
      submissions:
        description: Relations
        items:
          $ref: '#/definitions/entity.Submission'
        type: array
      tag:
        type: string
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      updated_at:
        type: string
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
...
  entity.RecentAction:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
  entity.Role:
    properties:
      created_at:
        type: string
      id:
        type: integer
      invites:
        items:
          $ref: '#/definitions/entity.Invite'
        type: array
      type:
        type: string
      updated_at:
        type: string
      users:
        description: Relations
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
  entity.Session:
    properties:
      attendances:
        description: Relations
        items:
          $ref: '#/definitions/entity.Attendance'
        type: array
      calendar_event_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      end_time:
        type: string
      fund:
        $ref: '#/definitions/entity.Fund'
      fund_id:
        type: integer
      group_sessions:
        items:
          $ref: '#/definitions/entity.GroupSession'
        type: array
      id:
        type: integer
      lecturer:
        $ref: '#/definitions/entity.User'
      lecturer_id:
        type: integer
      location:
        type: string
      meet_link:
        type: string
      name:
        type: string
      recording_link:
        type: string
      resource_link:
        type: string
      start_time:
        type: string
      stipends:
        items:
          $ref: '#/definitions/entity.Stipend'
        type: array
      updated_at:
        type: string
    type: object
  entity.Stipend:
    properties:
      created_at:
        type: string
      fund:
        $ref: '#/definitions/entity.Fund'
      fund_id:
        type: integer
      id:
        type: integer
      paid:
        type: boolean
      session:
        $ref: '#/definitions/entity.Session'
      session_id:
        type: integer
      share:
        description: Share of the fund amount
        type: number
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
  entity.Submission:
    properties:
      code:
        type: string
      comments:
        description: Relations
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      in_contest:
        description: Whether it was solved in a contest
        type: integer
      language:
        type: string
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      time_spent:
        description: Time spent in seconds
        type: integer
      tries:
        description: Number of attempts
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
      verified:
        type: boolean
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
...
  entity.SuperToGroup:
    properties:
      created_at:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      super_group:
        $ref: '#/definitions/entity.SuperGroup'
      super_group_id:
        type: integer
      updated_at:
        type: string
    type: object
  entity.Track:
    properties:
      active:
        type: boolean
      comments:
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      created_at:
        type: string
      exercises:
        items:
          $ref: '#/definitions/entity.Exercise'
        type: array
      id:
        type: integer
      name:
        type: string
      problem_tracks:
        items:
          $ref: '#/definitions/entity.ProblemTrack'
        type: array
      problems:
        description: Relations
        items:
          $ref: '#/definitions/entity.Problem'
        type: array
      super_group:
        $ref: '#/definitions/entity.SuperGroup'
      super_group_id:
        type: integer
      updated_at:
        type: string
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
  entity.User:
    properties:
      api_tokens:
        items:
          $ref: '#/definitions/entity.APIToken'
        type: array
      attendances:
        items:
          $ref: '#/definitions/entity.Attendance'
        type: array
      birthday:
        description: Personal Information
        type: string
      code_of_conduct:
        type: string
      codeforces:
        type: string
      comments:
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      config:
        type: string
      country:
        $ref: '#/definitions/entity.Country'
      country_id:
        type: integer
      created_at:
        description: Timestamps
        type: string
      cv:
        description: Professional Details
        type: string
      department:
        type: string
      email:
        type: string
      expected_graduation_date:
        type: string
      gender:
        type: string
      github:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      hackerrank:
        type: string
      head_attendances:
        items:
          $ref: '#/definitions/entity.Attendance'
        type: array
      id:
        description: Core Identity Fields
        type: integer
      inactive:
        type: boolean
      instagram:
        type: string
      joined_date:
        type: string
      leetcode:
        description: |-
          Coding Profiles (optional)
          Use pointer types so that if no value is provided, they remain nil.
        type: string
      linkedin:
        description: Social Media (all optional)
        type: string
      mentor_name:
        type: string
      name:
        type: string
      phone:
        description: Contact Information
        type: string
      photo:
        description: System Fields
        type: string
      posts:
        items:
          $ref: '#/definitions/entity.Post'
        type: array
      preferred_language:
        type: string
      recent_actions:
        items:
          $ref: '#/definitions/entity.RecentAction'
        type: array
      role:
        $ref: '#/definitions/entity.Role'
      role_id:
        description: Role and Relationships
        type: integer
      short_bio:
        type: string
      student_id:
        type: string
      submissions:
        description: Relations (using GORM associations)
        items:
          $ref: '#/definitions/entity.Submission'
        type: array
      telegram_uid:
        type: string
      telegram_username:
        type: string
      tshirt_color:
        description: Physical Attributes
        type: string
      tshirt_size:
        type: string
      university:
        description: Academic Information
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@hub.a2sv.org
    name: API Support
  description: This is the API documentation for the Hub backend.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Hub API
  version: "1.0"
paths:
...
  /api/sessions:
    get:
      description: Get a list of all sessions
      produces:
      - application/json
      responses:
        "200":
          description: List of sessions
          schema:
            items:
              $ref: '#/definitions/entity.Session'
            type: array
      summary: List sessions
      tags:
      - Sessions
    post:
      consumes:
      - application/json
      description: Create a new session for a user
      parameters:
      - description: Session data
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/entity.Session'
      produces:
      - application/json
      responses:
        "201":
          description: Session created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new session
      tags:
      - Sessions
  /api/sessions/{id}:
    delete:
      description: Delete a session by its ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Session deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid session ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties: true
            type: object
      summary: Delete a session
      tags:
      - Sessions
    get:
      description: Get a session by its ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Session details
          schema:
            $ref: '#/definitions/entity.Session'
        "400":
          description: Invalid session ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties: true
            type: object
      summary: Get session by ID
      tags:
      - Sessions
    patch:
      consumes:
      - application/json
      description: Update a session by its ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      - description: Session data
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/entity.Session'
      produces:
      - application/json
      responses:
        "200":
          description: Session updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties: true
            type: object
      summary: Update a session
      tags:
      - Sessions
  /api/sessions/name/{name}:
    get:
      description: Get a session by its name
      parameters:
      - description: Session name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session details
          schema:
            items:
              $ref: '#/definitions/entity.Session'
            type: array
      summary: Get session by name
      tags:
      - Sessions
  /api/sessions/start-time/{start_time}:
    get:
      description: Get a session by its start time
      parameters:
      - description: Session start time
        in: path
        name: start_time
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session details
          schema:
            items:
              $ref: '#/definitions/entity.Session'
            type: array
      summary: Get session by start time
      tags:
      - Sessions
  /api/stipends:
    get:
      description: Get a list of all stipends
      produces:
      - application/json
      responses:
        "200":
          description: List of stipends
          schema:
            items:
              $ref: '#/definitions/entity.Stipend'
            type: array
        "400":
          description: Failed to fetch stippends
          schema:
            additionalProperties: true
            type: object
      summary: List stipends
      tags:
      - Stipends
    post:
      consumes:
      - application/json
      description: Create a new stipend entry
      parameters:
      - description: Stipend data
        in: body
        name: stipend
        required: true
        schema:
          $ref: '#/definitions/entity.Stipend'
      produces:
      - application/json
      responses:
        "201":
          description: Stipend created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
      summary: Create a new stipend
      tags:
      - Stipends
  /api/stipends/{stippend_id}:
    delete:
      description: Delete a stipend by its ID
      parameters:
      - description: Stipend ID
        in: path
        name: stippend_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Stipend deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid stipend ID
          schema:
            additionalProperties: true
            type: object
      summary: Delete a stipend
      tags:
      - Stipends
    get:
      description: Get a stipend by its ID
      parameters:
      - description: Stipend ID
        in: path
        name: stippend_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Stipend details
          schema:
            $ref: '#/definitions/entity.Stipend'
        "400":
          description: Invalid stipend ID
          schema:
            additionalProperties: true
            type: object
      summary: Get stipend by ID
      tags:
      - Stipends
    patch:
      consumes:
      - application/json
      description: Update a stipend by its ID
      parameters:
      - description: Stipend ID
        in: path
        name: stippend_id
        required: true
        type: integer
      - description: Stipend data
        in: body
        name: stipend
        required: true
        schema:
          $ref: '#/definitions/entity.Stipend'
      produces:
      - application/json
      responses:
        "200":
          description: Stipend updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
      summary: Update a stipend
      tags:
      - Stipends
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"